tests pass

+++ interfaces/synthetix/IDepot.sol
@@ -1,3 +1,4 @@
+// SPDX-License-Identifier: You can't license an interface
 pragma solidity 0.6.8;
 
 
@@ -5,13 +6,16 @@ pragma solidity 0.6.8;
  * @title Synthetix Depot interface
  */
 interface IDepot {
-    function exchangeEtherForSynths() external payable returns (uint);
+    function exchangeEtherForSynths() external payable returns (uint256);
 
     // function exchangeEtherForSynthsAtRate(uint guaranteedRate) external payable returns (uint);
 
     // function depositSynths(uint amount) external;
 
-    function synthsReceivedForEther(uint amount) external view returns (uint);
+    function synthsReceivedForEther(uint256 amount)
+        external
+        view
+        returns (uint256);
 
     // function withdrawMyDepositedSynths() external;
 
@@ -24,20 +28,37 @@ interface IDepot {
     // Deprecated ABI for MAINNET. Only used on Testnets
     // function exchangeSynthsForSNX() external payable returns (uint);
 
-    event MaxEthPurchaseUpdated(uint amount);
+    event MaxEthPurchaseUpdated(uint256 amount);
     event FundsWalletUpdated(address newFundsWallet);
-    event Exchange(string fromCurrency, uint fromAmount, string toCurrency, uint toAmount);
-    event SynthWithdrawal(address user, uint amount);
-    event SynthDeposit(address indexed user, uint amount, uint indexed depositIndex);
-    event SynthDepositRemoved(address indexed user, uint amount, uint indexed depositIndex);
-    event SynthDepositNotAccepted(address user, uint amount, uint minimum);
-    event MinimumDepositAmountUpdated(uint amount);
-    event NonPayableContract(address indexed receiver, uint amount);
+    event Exchange(
+        string fromCurrency,
+        uint256 fromAmount,
+        string toCurrency,
+        uint256 toAmount
+    );
+    event SynthWithdrawal(address user, uint256 amount);
+    event SynthDeposit(
+        address indexed user,
+        uint256 amount,
+        uint256 indexed depositIndex
+    );
+    event SynthDepositRemoved(
+        address indexed user,
+        uint256 amount,
+        uint256 indexed depositIndex
+    );
+    event SynthDepositNotAccepted(
+        address user,
+        uint256 amount,
+        uint256 minimum
+    );
+    event MinimumDepositAmountUpdated(uint256 amount);
+    event NonPayableContract(address indexed receiver, uint256 amount);
     event ClearedDeposit(
         address indexed fromAddress,
         address indexed toAddress,
-        uint fromETHAmount,
-        uint toAmount,
-        uint indexed depositIndex
+        uint256 fromETHAmount,
+        uint256 toAmount,
+        uint256 indexed depositIndex
     );
 }
